solve recursion problem
previously using the same array and slicing out again and again.
as a result the array becomes too short at the end
and instead of returning the index from original array it was returning the index of sliced array
fix it with two extra parameter.
change the parameters without affecting the original array.